+++ tests/testthat/test-wrapper.R
@@ -5,6 +5,10 @@ context("wrapper")
 # changes in the operation model of biketown rather than
 # failure of the software
 biketown <- get_gbfs("portland")
+bike_itau <- get_gbfs("santiago")
+greenville_bcycle <- get_gbfs("greenville")
+veoride <- get_gbfs("https://share.veoride.com/api/share/gbfs")
+
 
 
 test_that("main wrapper works", {
@@ -12,6 +16,9 @@ test_that("main wrapper works", {
   skip_if_offline(host = "r-project.org")
   
   expect_equal(class(biketown), "list")
+  expect_equal(class(bike_itau), "list")
+  expect_equal(class(greenville_bcycle), "list")
+  expect_equal(class(veoride), "list")
   
 })